Fix AAC audio stuttering between ffmpeg runs
When transcoding, each new ffmpeg generates a file with a bit of silence
at the beginning to prime the AAC decoder. See
https://developer.apple.com/library/content/documentation/QuickTime/QTFF/QTFFAppenG/QTFFAppenG.html
for more information. These 1024 samples are audible as a slight
clicking noise in the stream.

The solution in this (very hacky and unfinished) commit is:
* Make audio segments exactly 4.992s or 324 AAC frames @ 48kHz long to
eliminate silence padding at the end
* Transcode one segment overlap for audio so that the segment containing
the silence does not ever need to be served.

This pretty much fixes the issue described in #1. However, there are
things left to do:

* Figure out if we really need the magic 4.992s or if ffmpeg will fill
up frames with audio anyway instead of padding
* Figure out if the magic 4.992s also works with other sampling rates
such as 44.1kHz.
* Clean up the messy code