Add Dockerfile with non-static ffmpeg in /opt/ffmpeg
Apparently building a static ffmpeg isn't that easy. By default
--enable static only makes the libs static it seems. See https://ffmpeg.org/pipermail/ffmpeg-user/2016-January/030278.html

I don't know a lot about linkers, but while experimenting with building
a static ffmpeg, it seemed to me like the linked libraries need to be
available in a format for static linking as well (and shared libraries
from Debian packages are not).