[streaming] Give out JWT for external subtitles
Giving out a filepath only works when the insecure allowDirectFileAccess
flag is enabled.

This is a consequence of the design decision of representing external
subtitles as their own file instead of as a "virtual stream" of the file
they belong to (in which case, the JWT the client already has could be
reused). Who knows, maybe it would have been smarter to do it the other
way and avoid having to sign tokens from the streaming server.