[streaming] Only allow one playback session per user/stream
This is to purge ffmpeg session that are probably no longer used because
the user reloaded the page.

In the future we probably want something smarter, like one stream per
file/device to better enable account sharing.

This change begs the question why we still have sessionID - if we end up
using the user ID as the primary key for the playback session anyway, we
may as well remove it.