Add mechanism to get a list of possible codecs before playing the file
That way, we can let the client check a list of codecs that we are able
to offer before requesting the manifest. When they request the manifest,
they give us back a list they can play and we can choose the best option
given the circumstances on the server. We do this on the server because

a) we have more information there than we would be able to put into a
standard DASH manifest. For example, we know which stream is transcoded.
b) this would put a lot of detail in the client, which we will possibly
implement more than once for different devices.