[metadata] Don't use GORM relations for PlayStates
Conceptually, they can fit into a has-many relationship pattern: A
Movie/Episode has-many PlayStates, but in practice, we always need to
filter by user ID, making it easier to just to the association by UUID
(not using the built-in mechanism) and just writing the association
manually.

This will allow us to simplify the API in the future.