Rate-limiting for missing metadata
Previously, the way this code worked was by spawning a new goroutine for
every single item with missing metadata. Probably fine if there are only
a few, but for larger libraries which may have hundreds or thousands of
such files, this was not great.

Now, this code limits the number of concurrent metadata update requests
based on the number of CPU cores detected.